Include windows cross-compile instructions in User Guide
The windows pull request mentioned on this page has been merged and is
now included in a tagged release, so we don't need to point to the pull
request.

There are just a few extra steps involved for a windows cross-compile;
these are set apart in 'note' sections which are rendered as visually
distinct 'note' boxes, communicating to the reader that they only apply
when cross-compiling for windows.